Solve Day 1, Part 2
I extracted the core parsing logic into an internal function that the
two main problem-solving functions delegate to.

I got hung up for a bit on how sort_by() mutates as opposed to returns a
new Vector. I'd like to improve this by finding a way to return a new
Vector and chain all of the operations.